+++ src/lib.rs
@@ -1,16 +1,8 @@
-// TODO: Everything here and in the pulseq-rs crate should use f64!
-// Individual samples can go down to sub-microsecond scale while sequence durations
-// often are in seconds or minutes. Single-precision f64 is not sufficient in the
-// general case. It probably is enough for simulation with absolute times, but
-// errors accumulate too quickly when generating sequences -> use double precision!
-
 mod backend_dsv;
 mod backend_pulseq;
-mod sequence;
 mod types;
 mod util;
 
-pub use sequence::Sequence;
 use std::path::Path;
 pub use types::*;
 
@@ -20,14 +12,64 @@ pub fn load_pulseq<P: AsRef<Path>>(path: P) -> Result<Sequence, pulseq_rs::Error
     )?)))
 }
 
-pub fn load_dsv<P: AsRef<Path>>(path: P, resolution: Option<usize>, ref_voltage: f64) -> Result<Sequence, backend_dsv::Error> {
-    Ok(Sequence(Box::new(backend_dsv::DsvSequence::load(path, resolution, ref_voltage)?)))
+pub fn load_dsv<P: AsRef<Path>>(
+    path: P,
+    resolution: Option<usize>,
+    ref_voltage: f64,
+) -> Result<Sequence, backend_dsv::Error> {
+    Ok(Sequence(Box::new(backend_dsv::DsvSequence::load(
+        path,
+        resolution,
+        ref_voltage,
+    )?)))
 }
 
-pub fn load_pulseq_str(source: &str) -> Result<Sequence, pulseq_rs::Error> {
-    Ok(Sequence(Box::new(
-        backend_pulseq::PulseqSequence::load_str(source)?,
-    )))
+/// A disseqt sequence. This opaque type on purpose does not expose the sequence data,
+/// but provides a simple interface which makes it possible to build importers and more
+/// that efficiently work with all supported MRI file formats.
+pub struct Sequence(pub(crate) Box<dyn Backend>);
+
+// Largely just forwards the trait impls, but also adds convenicence functions.
+impl Sequence {
+    pub fn fov(&self) -> Option<(f64, f64, f64)> {
+        self.0.fov()
+    }
+
+    pub fn duration(&self) -> f64 {
+        self.0.duration()
+    }
+    /// TODO: EventType should be the first parameter
+    pub fn encounter(&self, t_start: f64, ty: EventType) -> Option<(f64, f64)> {
+        self.0.encounter(t_start, ty)
+    }
+
+    pub fn events(&self, ty: EventType, t_start: f64, t_end: f64, max_count: usize) -> Vec<f64> {
+        self.0.events(ty, t_start, t_end, max_count)
+    }
+    /// TODO: EventType should be the first parameter
+    pub fn next_event(&self, t_start: f64, ty: EventType) -> Option<f64> {
+        self.events(ty, t_start, f64::INFINITY, 1).last().cloned()
+    }
+
+    pub fn sample(&self, time: &[f64]) -> SampleVec {
+        // TODO: We do a AoS -> SoA conversion here, which should be moved into the backend
+        // so the data can be emitted directly in the desired format
+        self.0.sample(time).into()
+    }
+
+    pub fn sample_one(&self, t: f64) -> Sample {
+        self.0.sample(&[t])[0]
+    }
+
+    pub fn integrate(&self, time: &[f64]) -> MomentVec {
+        // TODO: We do a AoS -> SoA conversion here, which should be moved into the backend
+        // so the data can be emitted directly in the desired format
+        self.0.integrate(time).into()
+    }
+
+    pub fn integrate_one(&self, t_start: f64, t_end: f64) -> Moment {
+        self.0.integrate(&[t_start, t_end])[0]
+    }
 }
 
 /// This trait is implemented by all backends and provides the basic functions
